pkg: binfmt: Workaround to EOVERFLOW for 32bit qemu on 64-bit host
readdir() returns NULL (errno=EOVERFLOW) for 32-bit user-static
qemu on 64-bit host. With this workaround we are using the qemu-arm
of Debian i386 packages, so we do not have the EOVERFLOW issue.

Check: https://bugs.launchpad.net/qemu/+bug/1805913

This can provide a solution to a specific issue, but it can also have
unknown side effects. More testing is needed.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>